Split files up, clean up, whitespace
