+++ Sources/BinaryJSON/Context.swift
@@ -11,71 +11,61 @@ import CLibbson
 public final class Context {
 
     // MARK: - Supporting Types
-    
-    public enum Option: UInt32, BitMaskOption {
-        
+
+    public enum Option: BitMaskOption {
+
         /// Context will be called from multiple threads.
-        case ThreadSafe
-        
+        case threadSafe
+
         /// Call ```getpid()``` instead of caching the result of ```getpid()``` when initializing the context.
-        case DisablePIDCache
-        
+        case disablePIDCache
+
         /// Call ```gethostname()``` instead of caching the result of ```gethostname()``` when initializing the context.
-        case DisableHostCache
-        
+        case disableHostCache
+
         //#if os(Linux)
         //case UseTaskID
         //#endif
-        
+
         public init?(rawValue: UInt32) {
-            
             switch rawValue {
-            case BSON_CONTEXT_THREAD_SAFE.rawValue: self = .ThreadSafe
-            case BSON_CONTEXT_DISABLE_PID_CACHE.rawValue: self = .DisablePIDCache
-            case BSON_CONTEXT_DISABLE_HOST_CACHE.rawValue: self = .DisableHostCache
-                //#if os(Linux)
+            case BSON_CONTEXT_THREAD_SAFE.rawValue: self = .threadSafe
+            case BSON_CONTEXT_DISABLE_PID_CACHE.rawValue: self = .disablePIDCache
+            case BSON_CONTEXT_DISABLE_HOST_CACHE.rawValue: self = .disableHostCache
+            //#if os(Linux)
                 //case BSON_CONTEXT_USE_TASK_ID.rawValue: self = .UseTaskID
             //#endif
             default: return nil
             }
         }
-        
+
         public var rawValue: UInt32 {
-            
             switch self {
-            case .ThreadSafe: return BSON_CONTEXT_THREAD_SAFE.rawValue
-            case .DisablePIDCache: return BSON_CONTEXT_DISABLE_PID_CACHE.rawValue
-            case .DisableHostCache: return BSON_CONTEXT_DISABLE_HOST_CACHE.rawValue
-                
-                //#if os(Linux)
+            case .threadSafe: return BSON_CONTEXT_THREAD_SAFE.rawValue
+            case .disablePIDCache: return BSON_CONTEXT_DISABLE_PID_CACHE.rawValue
+            case .disableHostCache: return BSON_CONTEXT_DISABLE_HOST_CACHE.rawValue
+            //#if os(Linux)
                 //case .UseTaskID: return BSON_CONTEXT_USE_TASK_ID.rawValue
-                //#endif
+            //#endif
             }
         }
     }
 
-    /// The default, thread-safe BSON context.
-    public static let defualt = Context(options: [.ThreadSafe, .DisablePIDCache])
-    
-    // MARK: - Properties
-    
+    public static let `default` = Context(options: [.threadSafe, .disablePIDCache])
+
     public let options: [Option]
-    
-    // MARK: - Internal Properties
-    
     internal let internalPointer: OpaquePointer
-    
-    // MARK: - Initialization
-    
-    deinit { bson_context_destroy(internalPointer) }
-    
+
     /// Initializes the context with the specified options.
     public init(options: [Option] = []) {
-        
         let flags = options.optionsBitmask()
-        
+
         self.options = options
         self.internalPointer = bson_context_new(bson_context_flags_t(rawValue: flags))
+    }
+
+    deinit {
+        bson_context_destroy(internalPointer)
     }
 }
