+++ Sources/BinaryJSON/Key.swift
@@ -1,0 +1,13 @@
+//
+//  Key.swift
+//  BSON
+//
+//  Created by Dan Appel on 06/27/16.
+//  Copyright © 2016 Dan Appel. All rights reserved.
+//
+
+/// BSON maximum and minimum representable types.
+public enum Key {
+    case min
+    case max
+}
