+++ Sources/BinaryJSON/Reader.swift
@@ -9,22 +9,17 @@
 import CLibbson
 
 public final class Reader: IteratorProtocol {
-    
-    // MARK: - Private Properties
-    
+
     private let internalPointer: UnsafeMutablePointer<bson_reader_t>
-    
-    // MARK: - Initialization
-    
+
     public init(data: Data) {
-        
         self.internalPointer = bson_reader_new_from_data(data.bytes, data.bytes.count)
     }
-    
-    deinit { bson_reader_destroy(internalPointer) }
-    
-    // MARK: - Methods
-    
+
+    deinit {
+        bson_reader_destroy(internalPointer)
+    }
+
     public func next() -> [String:BSON]? {
         var eof = false
